+++ viadot/sources/sap_rfc.py
@@ -1,4 +1,5 @@
 import re
+from prefect.utilities import logging
 from collections import OrderedDict
 from typing import List, Literal
 from typing import OrderedDict as OrderedDictType
@@ -8,13 +9,16 @@ import pandas as pd
 
 try:
     import pyrfc
+    from pyrfc._exception import ABAPApplicationError
 except ModuleNotFoundError:
     raise ImportError("pyfrc is required to use the SAPRFC source.")
 from sql_metadata import Parser
 
 from viadot.config import local_config
-from viadot.exceptions import CredentialError
+from viadot.exceptions import CredentialError, DataBufferExceeded
 from viadot.sources.base import Source
+
+logger = logging.get_logger()
 
 
 def remove_whitespaces(text):
@@ -358,6 +362,26 @@ class SAPRFC(Source):
         # this has to be called before checking client_side_filters
         where = self.where
         columns = self.select_columns
+        # due to the RFC_READ_TABLE limit of characters per row, colums are splited into smaller lists
+        lists_of_columns = []
+        cols = []
+        col_length_total = 0
+        for col in columns:
+            info = self.call("DDIF_FIELDINFO_GET", TABNAME=table_name, FIELDNAME=col)
+            col_length = info["DFIES_TAB"][0]["LENG"]
+            col_length_total += int(col_length)
+            # According to SAP documentation, the limit is 512 characters. However, we observed SAP raising an exception even
+            # on a slightly lower number of characters, so we add a safety margin
+            character_limit = 400
+            if col_length_total <= character_limit:
+                cols.append(col)
+            else:
+                lists_of_columns.append(cols)
+                cols = [col]
+                col_length_total = 0
+        lists_of_columns.append(cols)
+
+        columns = lists_of_columns
         options = [{"TEXT": where}] if where else None
         limit = self._get_limit(sql)
         offset = self._get_offset(sql)
@@ -393,7 +417,8 @@ class SAPRFC(Source):
         characters, we trim whe WHERE clause and perform the rest of the filtering
         on the resulting DataFrame. Eg. if the WHERE clause contains 4 conditions
         and has 80 characters, we only perform 3 filters in the query, and perform
-        the last filter on the DataFrame.
+        the last filter on the DataFrame. If characters per row limit will be exceeded,
+        data will be downloaded in chunks.
 
         Source: https://success.jitterbit.com/display/DOC/Guide+to+Using+RFC_READ_TABLE+to+Query+SAP+Tables#GuidetoUsingRFC_READ_TABLEtoQuerySAPTables-create-the-operation
         - WHERE clause: 75 character limit
@@ -405,6 +430,9 @@ class SAPRFC(Source):
         params = self._query
         columns = self.select_columns_aliased
         sep = self._query.get("DELIMITER")
+        fields_lists = self._query.get("FIELDS")
+        if len(fields_lists) > 1:
+            logger.info(f"Data will be downloaded in {len(fields_lists)} chunks.")
         func = self.func
         if sep is None:
             # automatically find a working separator
@@ -427,18 +455,35 @@ class SAPRFC(Source):
 
         records = None
         for sep in SEPARATORS:
+            logger.info(f"Checking if separator '{sep}' works.")
+            df = pd.DataFrame()
             self._query["DELIMITER"] = sep
-            try:
-                response = self.call(func, **params)
-                record_key = "WA"
-                data_raw = response["DATA"]
-                records = [row[record_key].split(sep) for row in data_raw]
-            except ValueError:
-                continue
+            chunk = 1
+            for fields in fields_lists:
+                logger.info(f"Downloading {chunk} data chunk...")
+                try:
+                    self._query["FIELDS"] = fields
+                    try:
+                        response = self.call(func, **params)
+                    except ABAPApplicationError as e:
+                        if e.key == "DATA_BUFFER_EXCEEDED":
+                            raise DataBufferExceeded(
+                                "Character limit per row exceeded. Please select fewer columns."
+                            )
+                        else:
+                            raise e
+                    record_key = "WA"
+                    data_raw = response["DATA"]
+                    records = [row[record_key].split(sep) for row in data_raw]
+                    df[fields] = records
+                    chunk += 1
+                except ValueError:
+                    df = pd.DataFrame()
+                    continue
         if records is None:
             raise ValueError("None of the separators worked.")
 
-        df = pd.DataFrame(records, columns=columns)
+        df.columns = columns
 
         if self.client_side_filters:
             filter_query = self._build_pandas_filter_query(self.client_side_filters)